requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
attrs==21.4.0
bcrypt==3.2.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
ConfigArgParse==1.5.3
contextlib2==21.6.0
contextvars==2.4
cryptography==36.0.1
dataclasses==0.8
Deprecated==1.2.13
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.3
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
locust==2.5.1
Mako==1.1.6
MarkupSafe==2.0.1
msgpack==1.0.3
orjson==3.6.1
packaging==21.3
passlib==1.7.4
pkg_resources==0.0.0
pluggy==1.0.0
psutil==5.9.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
pyzmq==22.3.0
redis==4.1.0
requests==2.26.0
roundrobin==0.0.2
rsa==4.8
simple-crypt==4.1.7
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.0
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.0
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
